Lesson 3/requirements.txt
python-telegram-bot==22.7
requests==2.31.0
pydantic-settings==2.7.1
